+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 0.22.3 (2026-05-01)
 
 - Change value of `Path::MAX_COMPONENT_LENGTH` to `u16::MAX - 2` [#3087](https://github.com/0xMiden/miden-vm/pull/3087)
+- Optimized range check lookup tracking to avoid heap allocations during trace generation ([#2793](https://github.com/0xMiden/miden-vm/issues/2793)).
 
 ## 0.22.2 (2026-04-028)